+++ srcs/requirements/django/src/game/routing.py
@@ -1,11 +1,11 @@
 from django.urls import re_path
-from . import matching
+from . import pongConsumers
 from django.urls import path
 from game.consumers import practicePongConsumers, practiceFightingConsumers, fightConsumers
 
 websocket_urlpatterns = [
-	re_path("ws/game/pongonebyone", matching.PongOneMatch.as_asgi()),
-	# re_path("ws/game/pongtwobytwo", matching.PongTwoMatch.as_asgi()),
-	# re_path("ws/game/pongtournament", matching.PongTournamentMatch.as_asgi()),
+	re_path("ws/game/pongonebyone", pongConsumers.PongOneMatch.as_asgi()),
+	re_path("ws/game/pongtwobytwo", pongConsumers.PongTwoMatch.as_asgi()),
+	re_path("ws/game/pongtournament", pongConsumers.PongTournamentMatch.as_asgi()),
 	path("ws/game/fighting", fightConsumers.fightingConsumers.as_asgi()),
 ]
